refactor(web): dirs structure and other improvements
