+++ webdriver-ts-results/src/App.css
@@ -25,32 +25,6 @@ select {
   font-size: 14px;
 }
 
-.select-bar {
-  background-color: #fafafa;
-  border-top: 1px solid var(--border-color);
-  border-bottom: 1px solid var(--border-color);
-  padding: 10px 0;
-  line-height: 0;
-}
-.select-bar__dropdowns {
-  padding: 5px 0;
-  display: flex;
-  flex-direction: row;
-  gap: 1rem;
-}
-
-@media screen and (width>=640px) {
-  .select-bar {
-    flex-direction: unset;
-    align-items: center;
-  }
-}
-@media screen and (width>=768px) {
-  .select-bar {
-    align-items: center;
-  }
-}
-
 select {
   border: 1px solid var(--border-color);
   border-radius: 4px;
@@ -62,10 +36,6 @@ input + label {
   margin-left: 5px;
 }
 
-.panel-body .checkbox {
-  margin-top: 5px;
-  margin-bottom: 5px;
-}
 thead.dummy th {
   padding: 0;
   border: 0;
