+++ webdriver-ts-results/src/assets/icons/index.ts
@@ -1,0 +1,4 @@
+import CopyIcon from "./CopyIcon.tsx";
+import PasteIcon from "./PasteIcon.tsx";
+
+export { CopyIcon, PasteIcon };
