--- webdriver-ts-results/src/components/selection/SelectionBar.tsx
@@ -1,27 +1,0 @@
-import BenchmarkSelector from "./BenchmarkSelector";
-import FrameworkSelector from "./FrameworkSelector";
-import ModeSelecionPanel from "./ModeSelectionPanel";
-import CopyPasteSelection from "./CopyPasteSelection";
-
-interface Props {
-  showDurationSelection: boolean;
-}
-
-export const SelectionBar = ({ showDurationSelection }: Props) => {
-  console.log("SelectionBar");
-
-  return (
-    <>
-      <div className="select-bar">
-        <div className="select-bar__dropdowns">
-          <FrameworkSelector />
-          <BenchmarkSelector />
-          <CopyPasteSelection />
-        </div>
-        <div className="select-bar__dropdowns">
-          <ModeSelecionPanel showDurationSelection={showDurationSelection} />
-        </div>
-      </div>
-    </>
-  );
-};
