+++ webdriver-ts-results/src/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import App from "./App";
+import App from "./App.tsx";
 import { createRoot } from "react-dom/client";
 
 createRoot(document.getElementById("root")!).render(
